+++ packages/matrix-code/src/components/molecules/forge-matrix-code/forge-matrix-code.tsx
@@ -2,7 +2,7 @@ import { ForgeButton } from '@mission-platform/components';
 import { ForgeIconCheck, ForgeIconCopy, ForgeIconDownload, ForgeIconImage } from '@mission-platform/icons';
 import { h, type MpElement, useEffect, useMemo, useRef, useState } from '@mission-platform/forge';
 import { ForgeTypography } from '@mission-platform/typography';
-import { encodeMatrix, type MatrixSymbology } from '@/encoder';
+import { encodeMatrix, type MatrixSymbology } from '@mission-platform/matrix-code';
 
 import styles from './forge-matrix-code.module.scss';